+++ source/Core/Utils/GASSTimer.cpp
@@ -48,7 +48,7 @@ namespace GASS
 			QueryPerformanceCounter((LARGE_INTEGER *) &m_PerfTimerStart);
 			m_PerfTimer		= true;				// Set Performance Timer To TRUE
 			// Calculate The Timer Resolution Using The Timer Frequency
-			m_Resolution		= (float) (((double)1.0f)/((double)m_Frequency));
+			m_Resolution		= static_cast<float>(1.0/static_cast<double>(m_Frequency));
 			// Set The Elapsed Time To The Current Time
 			m_PerfTimerElapsed	= m_PerfTimerStart;
 		}
@@ -84,12 +84,12 @@ namespace GASS
 		{
 			QueryPerformanceCounter((LARGE_INTEGER *) &time);		// Grab The Current Performance Time
 			// Return The Current Time Minus The Start Time Multiplied By The Resolution
-			return ( (double) ( time - m_PerfTimerStart) * m_Resolution);
+			return static_cast<double>( time - m_PerfTimerStart) * m_Resolution;
 		}
 		else
 		{
 			// Return The Current Time Minus The Start Time Multiplied By The Resolution
-			return( (double) ( timeGetTime() - m_MMTimerStart) * m_Resolution);
+			return static_cast<double>( timeGetTime() - m_MMTimerStart) * m_Resolution;
 		}
 
 #else
